pyproject.toml
[tool.poetry]
name = "pybdx"
version = "0.2.0"
description = "BDX XML data fetcher and formatter"
authors = ["Joey Grable <hello@joeygrable.com>"]
license = "GNU GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
autopep8 = "1.6.0"
bcrypt = "3.2.0"
beautifulsoup4 = "4.10.0"
cffi = "^1.15.1"
cryptography = "^40.0.2"
lxml = "^4.9.2"
numpy = "^1.24.3"
pandas = "^2.0.1"
paramiko = "^3.1.0"
pep8 = "^1.7.1"
phpserialize = "^1.3"
pycodestyle = "^2.10.0"
pycparser = "^2.21"
pynacl = "^1.5.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
six = "^1.16.0"
soupsieve = "^2.4.1"
toml = "^0.10.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
